refactor!: address PR review feedback for annotations migration
- Remove #ifdef RCT_NEW_ARCH_ENABLED guards from iOS ComponentViews
- Delete old arch iOS managers (MLRNPointAnnotationManager, MLRNCalloutManager)
- Fix PointAnnotation onDeselected not firing on Android map click
- Remove unimplemented allowOverlap and isSelected from MarkerView native spec
- Restructure annotations into per-component subdirectories (point-annotation/, marker-view/, callout/)

BREAKING CHANGE: Annotation native files moved to subdirectories.
Android packages changed from annotations to annotations.pointannotation,
annotations.markerview, and annotations.callout.
