+++ android/src/main/java/org/maplibre/reactnative/components/annotations/callout/MLRNCallout.kt
@@ -1,4 +1,4 @@
-package org.maplibre.reactnative.components.annotations
+package org.maplibre.reactnative.components.annotations.callout
 
 import android.content.Context
 import com.facebook.react.views.view.ReactViewGroup
