+++ android/src/main/java/org/maplibre/reactnative/components/annotations/callout/MLRNCalloutManager.kt
@@ -1,4 +1,4 @@
-package org.maplibre.reactnative.components.annotations
+package org.maplibre.reactnative.components.annotations.callout
 
 import com.facebook.react.module.annotations.ReactModule
 import com.facebook.react.uimanager.ThemedReactContext
