--- ios/components/annotations/MLRNCalloutManager.h
@@ -1,5 +1,0 @@
-#import "ViewManager.h"
-
-@interface MLRNCalloutManager : ViewManager
-
-@end
