--- ios/components/annotations/MLRNCalloutManager.m
@@ -1,12 +1,0 @@
-#import "MLRNCalloutManager.h"
-#import "MLRNCallout.h"
-
-@implementation MLRNCalloutManager
-
-RCT_EXPORT_MODULE()
-
-- (UIView *)view {
-  return [[MLRNCallout alloc] init];
-}
-
-@end
