--- ios/components/annotations/MLRNPointAnnotationManager.h
@@ -1,5 +1,0 @@
-#import "ViewManager.h"
-
-@interface MLRNPointAnnotationManager : ViewManager
-
-@end
