+++ ios/components/annotations/callout/MLRNCalloutComponentView.h
@@ -1,5 +1,3 @@
-#ifdef RCT_NEW_ARCH_ENABLED
-
 #import <React/RCTViewComponentView.h>
 #import <UIKit/UIKit.h>
 
@@ -10,5 +8,3 @@ NS_ASSUME_NONNULL_BEGIN
 @end
 
 NS_ASSUME_NONNULL_END
-
-#endif
