+++ src/utils/animated/Animated.tsx
@@ -5,7 +5,7 @@ import { AnimatedCoordinatesArray } from "./AnimatedCoordinatesArray";
 import { AnimatedExtractCoordinateFromArray } from "./AnimatedExtractCoordinateFromArray";
 import { AnimatedGeoJSON } from "./AnimatedGeoJSON";
 import { AnimatedRouteCoordinatesArray } from "./AnimatedRouteCoordinatesArray";
-import { MarkerView } from "../../components/annotations/MarkerView";
+import { MarkerView } from "../../components/annotations/marker-view/MarkerView";
 import { Layer } from "../../components/layers/Layer";
 import { GeoJSONSource } from "../../components/sources/geojson-source/GeoJSONSource";
 import { ImageSource } from "../../components/sources/image-source/ImageSource";
